+++ R/XStARS.R
@@ -82,7 +82,7 @@
 #' @examples
 #' cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
 #' \donttest{
-#' # Takes around 20 seconds
+#' # Takes around one minute
 #' sel_cg <- xstars(cg, verbose = FALSE)
 #' }
 xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NULL, rep_num = 20, max_iter = 10, verbose = TRUE) {
@@ -111,7 +111,7 @@ xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NUL
   
   for (i in 1:n_c) {
     if (verbose) {
-      mes <- paste(c("Selecting best Lambda_w/Lambda_b combination for all c values with \"stars\"....in progress:", floor(100 * i / n_c), "%"), collapse = "")
+      mes <- paste(c("Selecting best Lambda_w/Lambda_b combination for all c values with \"xstars\"....in progress:", floor(100 * i / n_c), "%"), collapse = "")
       cat(mes, "\r")
       cat("\n")
       flush.console()
@@ -141,7 +141,10 @@ xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NUL
           }
           ind.sample <- sample(c(1:n), floor(n * stars_subsample_ratio), replace = FALSE)
           
-          tmp <- coglasso(coglasso_obj$data[ind.sample, ], pX = pX, lambda_w = coglasso_obj$lambda_w, lambda_b = c(lb_sel), c = c(coglasso_obj$c[i]), verbose = FALSE)
+          corr_matrix <- cor(scale(coglasso_obj$data[ind.sample, ]))
+          hpars <- matrix(c(rep(alpha, n_lambda_w), coglasso_obj$lambda_w, rep(lb_sel, n_lambda_w)), nrow = n_lambda_w, ncol = 3)
+          tmp <- co_glasso(corr_matrix, pX, hpars, FALSE, FALSE, FALSE)
+          
           convergence <- tmp$convergence
           tmp <- tmp$path
           
@@ -196,7 +199,10 @@ xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NUL
           }
           ind.sample <- sample(c(1:n), floor(n * stars_subsample_ratio), replace = FALSE)
           
-          tmp <- coglasso(coglasso_obj$data[ind.sample, ], pX = pX, lambda_w = c(lw_sel), lambda_b = coglasso_obj$lambda_b, c = c(coglasso_obj$c[i]), verbose = FALSE)
+          corr_matrix <- cor(scale(coglasso_obj$data[ind.sample, ]))
+          hpars <- matrix(c(rep(alpha, n_lambda_b), coglasso_obj$lambda_b, rep(lw_sel, n_lambda_b)), nrow = n_lambda_b, ncol = 3)
+          tmp <- co_glasso(corr_matrix, pX, hpars, FALSE, FALSE, FALSE)
+          
           convergence <- tmp$convergence
           tmp <- tmp$path
           
@@ -252,7 +258,413 @@ xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NUL
     
     
     if (verbose) {
-      mes <- "Selecting best Lambda_w/Lambda_b combination for all c values with \"stars\"....done"
+      mes <- "Selecting best Lambda_w/Lambda_b combination for all c values with \"xstars\"....done"
+      cat(mes, "\r")
+      flush.console()
+    }
+    
+    coglasso_obj$opt_index_lw[i] <- index_lw
+    coglasso_obj$opt_index_lb[i] <- index_lb
+    coglasso_obj$opt_lambda_w[i] <- coglasso_obj$lambda_w[index_lw]
+    coglasso_obj$opt_lambda_b[i] <- coglasso_obj$lambda_b[index_lb]
+    opt_hpars_combination <- which(coglasso_obj$hpars[, 1] == alpha &
+                                     coglasso_obj$hpars[, 2] == coglasso_obj$opt_lambda_w[i] &
+                                     coglasso_obj$hpars[, 3] == coglasso_obj$opt_lambda_b[i])
+    coglasso_obj$opt_adj[[i]] <- coglasso_obj$path[[opt_hpars_combination]]
+    colnames(coglasso_obj$opt_adj[[i]]) <- colnames(coglasso_obj$data)
+    row.names(coglasso_obj$opt_adj[[i]]) <- colnames(coglasso_obj$data)
+  }
+  
+  if (all(is.na(coglasso_obj$opt_variability))) {
+    warning("coglasso did not converge for any c parameter. Will select the highest c.")
+  }
+  
+  coglasso_obj$sel_index_c <- min(which.min(coglasso_obj$opt_variability), n_c)
+  coglasso_obj$sel_index_lw <- coglasso_obj$opt_index_lw[coglasso_obj$sel_index_c]
+  coglasso_obj$sel_index_lb <- coglasso_obj$opt_index_lb[coglasso_obj$sel_index_c]
+  coglasso_obj$sel_lambda_w <- coglasso_obj$lambda_w[coglasso_obj$sel_index_lw]
+  coglasso_obj$sel_lambda_b <- coglasso_obj$lambda_b[coglasso_obj$sel_index_lb]
+  coglasso_obj$sel_c <- coglasso_obj$c[coglasso_obj$sel_index_c]
+  coglasso_obj$sel_adj <- coglasso_obj$opt_adj[[coglasso_obj$sel_index_c]]
+  sel_hpars_combination <- which(coglasso_obj$hpars[, 1] == 1 / (1 + coglasso_obj$sel_c) &
+                                   coglasso_obj$hpars[, 2] == coglasso_obj$sel_lambda_w &
+                                   coglasso_obj$hpars[, 3] == coglasso_obj$sel_lambda_b)
+  coglasso_obj$sel_density <- coglasso_obj$density[sel_hpars_combination]
+  coglasso_obj$sel_icov <- coglasso_obj$icov[[sel_hpars_combination]]
+  if (!is.null(coglasso_obj$cov)) {
+    coglasso_obj$sel_cov <- coglasso_obj$cov[[sel_hpars_combination]]
+  }
+  
+  return(coglasso_obj)
+}
+
+#' Efficient stability selection of the best `coglasso` network
+#'
+#' @description `xestars()` provides a more efficient and lighter implementation
+#' than `xstars()` to select the combination of hyperparameters given to
+#' `coglasso()` yielding the most stable, yet sparse network. Stability is
+#' computed upon network estimation from multiple subsamples of the multi-omics
+#' data set, allowing repetition. Subsamples are collected for a fixed amount of
+#' times (`rep_num`), and with a fixed proportion of the total number of samples
+#' (`stars_subsample_ratio`).
+#'
+#' @details
+#' *eXtended Efficient StARS* (*XEStARS*) is a more efficient and memory-light version of
+#' *XStARS*, the adaptation for *collaborative graphical regression* of the method
+#' published by Liu, H. *et al.* (2010): Stability Approach to Regularization
+#' Selection (StARS). *StARS* was developed for network estimation regulated by
+#' a single penalty parameter, while *collaborative graphical lasso* needs to
+#' explore three different hyperparameters. In particular, two of these are
+#' penalty parameters with a direct influence on network sparsity, hence on
+#' stability. For every \eqn{c} parameter, `xestars()` explores one of the two
+#' penalty parameters (\eqn{\lambda_w} or \eqn{\lambda_b}), keeping the other
+#' one fixed at its previous best estimate, using the normal, one-dimentional
+#' *StARS* approach, until finding the best couple. What makes it more efficient
+#' than `xstars()` is that the stability check that in the original algorithm 
+#' (even in the original *StARS*) is performed for every \eqn{\lambda_w} or 
+#' \eqn{\lambda_b} value, is implemented here as a *stopping criterion*. This 
+#' reduces sensibly the number of iterations before convergence. It then selects
+#' the \eqn{c} parameter for which the best (\eqn{\lambda_w}, \eqn{\lambda_b})
+#' couple yielded the most stable, yet sparse network. \cr
+#' The original *XStARS* computes a new subsampling for every time the algorithm
+#' switches from optimizing the two\eqn{\lambda_w} and \eqn{\lambda_b}, and for 
+#' every \eqn{c}. This does not allow to compare the hyperparameters on an equal
+#' ground, and can slow the selection down with bigger data set or a larger
+#' hyperparameter space. To allow a fairer (and faster) comparison among 
+#' different optimizations, the `old_sampling` parameter has been implemented. 
+#' If set to TRUE, the subsampling is the same one `xstars()` would perform. 
+#' Otherwise the subsampling is performed at the beginning of the algorithm once
+#' and for all its iterations. \cr
+#' To allow `xestars()` to be more memory light, the `light` parameter has been 
+#' implemented. If set to TRUE and the "merged" matrixes traditionally returned 
+#' by both *StARS* and *XStARS* are not returned.
+#' 
+#' @param coglasso_obj The object returned by `coglasso()`.
+#' @param stars_thresh The threshold set for variability of the explored
+#'   networks at each iteration of the algorithm. The \eqn{\lambda_w} or the
+#'   \eqn{\lambda_b} associated to the most stable network before the threshold
+#'   is overcome is selected.
+#' @param stars_subsample_ratio The proportion of samples in the multi-omics
+#'   data set to be randomly subsampled to estimate the variability of the
+#'   network under the given hyperparameters setting. Defaults to 80% when the
+#'   number of samples is smaller than 144, otherwise it defaults to
+#'   \eqn{\frac{10}{n}\sqrt{n}}.
+#' @param rep_num The amount of subsamples of the multi-omics data set used to
+#'   estimate the variability of the network under the given hyperparameters
+#'   setting. Defaults to 20.
+#' @param max_iter The greatest number of times the algorithm is allowed to
+#'   choose a new best \eqn{\lambda_w}. Defaults to 10.
+#' @param old_sampling If set to TRUE `xestars()` performs the same subsampling 
+#'   `xstars()` would. Makes a difference with bigger data sets, where computing
+#'   a correlation matrix could take significantly longer. Defaults to FALSE.
+#' @param light If set TRUE the "merged" matrixes recording average variability
+#'   of each edge are not stored and the algorithm is more memory efficient. 
+#'   Defaults to TRUE.
+#' @param verbose Print information regarding the progress of the selection
+#'   procedure on the console.
+#'
+#' @return `xestars()` returns a list containing the results of the selection
+#'   procedure, built upon the list returned by `coglasso()`.
+#' * ... are the same elements returned by [coglasso()].
+#' * `opt_adj` is a list of the adjacency matrices finally selected for each
+#'   \eqn{c} parameter explored.
+#' * `opt_variability` is a numerical vector containing the variabilities
+#'   associated to the adjacency matrices in `opt_adj`.
+#' * `opt_index_lw` and `opt_index_lb` are integer vectors containing the
+#'   index of the selected \eqn{\lambda_w}s (or \eqn{\lambda_b}s) for each
+#'   \eqn{c} parameters explored.
+#' * `opt_lambda_w` and `opt_lambda_b` are vectors containing the selected
+#'   \eqn{\lambda_w}s (or \eqn{\lambda_b}s) for each \eqn{c} parameters
+#'   explored.
+#' * `sel_index_c`, `sel_index_lw` and `sel_index_lb` are the indexes of the
+#'   final selected parameters \eqn{c}, \eqn{\lambda_w} and \eqn{\lambda_b}
+#'   leading to the most stable sparse network.
+#' * `sel_c`, `sel_lambda_w` and `sel_lambda_b` are the final selected
+#'   parameters \eqn{c}, \eqn{\lambda_w} and \eqn{\lambda_b} leading to the most
+#'   stable sparse network.
+#' * `sel_adj` is the adjacency matrix of the final selected network.
+#' * `sel_density` is the density of the final selected network.
+#' * `sel_icov` is the inverse covariance matrix of the final selected network.
+#' * `merge_lw` and `merge_lb` are returned only if `light` is set to FALSE.
+#'   They are lists with as many elements as the number of
+#'   \eqn{c} parameters explored. Every element is a "merged" adjacency matrix,
+#'   the average of all the adjacency matrices estimated  for those specific 
+#'   \eqn{c} and the selected \eqn{\lambda_w} (or \eqn{\lambda_b}) values 
+#'   across all the subsampling in the last path explored before convergence, 
+#'   the one when the final combination of \eqn{\lambda_w} and \eqn{\lambda_b} 
+#'   is selected for the given \eqn{c} value.
+#'
+#' @export
+#'
+#' @examples
+#' cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
+#' \donttest{
+#' # Takes less than five seconds
+#' sel_cg <- xestars(cg, verbose = FALSE)
+#' }
+xestars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NULL, rep_num = 20, max_iter = 10, old_sampling = FALSE, light = TRUE, verbose = TRUE) {
+  n <- nrow(coglasso_obj$data)
+  d <- ncol(coglasso_obj$data)
+  pX <- coglasso_obj$pX
+  n_lambda_w <- length(coglasso_obj$lambda_w)
+  n_lambda_b <- length(coglasso_obj$lambda_b)
+  n_c <- length(coglasso_obj$c)
+  
+  if (is.null(stars_subsample_ratio)) {
+    if (n > 144) stars_subsample_ratio <- 10 * sqrt(n) / n
+    if (n <= 144) stars_subsample_ratio <- 0.8
+  }
+  
+  if (!light) {
+    coglasso_obj$merge_lw <- vector(mode = "list", length = n_c)
+    coglasso_obj$merge_lb <- vector(mode = "list", length = n_c)
+  }
+  
+  coglasso_obj$opt_adj <- vector(mode = "list", length = n_c)
+  coglasso_obj$opt_variability <- rep(0, n_c)
+  coglasso_obj$opt_index_lw <- rep(0, n_c)
+  coglasso_obj$opt_index_lb <- rep(0, n_c)
+  coglasso_obj$opt_lambda_w <- rep(0, n_c)
+  coglasso_obj$opt_lambda_b <- rep(0, n_c)
+  
+  # Create the same samples of indexes for the whole set of iterations,
+  # cancelled when `old_sampling` is true
+  if (!old_sampling) {
+    corr_matrixes <- vector(mode = "list", length = rep_num)
+    for (j in 1:rep_num) {
+      ind.sample <- sample(c(1:n), floor(n * stars_subsample_ratio), replace = FALSE)
+      corr_matrixes[[j]] <- cor(scale(coglasso_obj$data[ind.sample, ]))
+    }
+  }
+  
+  for (i in 1:n_c) {
+    if (verbose) {
+      mes <- paste(c("Selecting best Lambda_w/Lambda_b combination for all c values with \"xestars\"....in progress:", floor(100 * i / n_c), "%"), collapse = "")
+      cat(mes, "\r")
+      cat("\n")
+      flush.console()
+    }
+    
+    alpha <- 1 / (1 + coglasso_obj$c[i])
+    
+    lw_sel <- -1
+    lb_sel <- coglasso_obj$lambda_b[n_lambda_b]
+    num_iter <- 0
+    converged <- FALSE
+    select_lw <- TRUE
+    
+    while (!converged & num_iter < max_iter) {
+      if (select_lw) {
+        # Consider breaking in a one-dimensional sub-function returning a lambda estars
+        
+        if (old_sampling) {
+          corr_matrixes <- vector(mode = "list", length = rep_num)
+          for (j in 1:rep_num) {
+            ind.sample <- sample(c(1:n), floor(n * stars_subsample_ratio), replace = FALSE)
+            corr_matrixes[[j]] <- cor(scale(coglasso_obj$data[ind.sample, ]))
+          }
+        }
+        
+        for (j in 1:n_lambda_w) {
+          if (verbose) {
+            mes <- paste(c("Conducting Subsampling Lambda_w....in progress:", floor(100 * j / n_lambda_w), "%"), collapse = "")
+            cat(mes, "\r")
+            flush.console()
+          }
+          real_rep.num <- 0
+          merge_tmp <- rep(0, d*d)
+          for (k in 1:rep_num) {
+            tmp <- co_glasso(corr_matrixes[[k]], pX, t(as.matrix(c(alpha, coglasso_obj$lambda_w[j], lb_sel))), FALSE, FALSE, FALSE)
+            convergence <- tmp$convergence[1]
+            tmp <- as.vector(tmp$path[[1]])
+            if (convergence == 1) {
+              real_rep.num <- real_rep.num + 1
+              merge_tmp <- merge_tmp + tmp
+            }
+          }
+          rm(tmp)
+          gc()
+          
+          merge_tmp <- merge_tmp / real_rep.num
+          variability_tmp <- 4 * sum(merge_tmp * (1 - merge_tmp)) / (d * (d - 1))
+          
+          # Find way to deal with when real_rep.num is 0 (no convergence at all)
+          if (variability_tmp >= stars_thresh) {
+            index_lw <- max(j - 1, 1)
+            tmp_lw <- coglasso_obj$lambda_w[[index_lw]]
+            if (tmp_lw == lw_sel) converged <- TRUE
+            
+            lw_sel <- tmp_lw
+            
+            if (j == 1) {
+              coglasso_obj$opt_variability[i] <- variability_tmp
+              if (!light) {
+                coglasso_obj$merge_lw[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+              }
+            }
+            
+            if (verbose) {
+              mes <- "Conducting Subsampling Lambda_w....done.                 "
+              cat(mes, "\r")
+              cat("\n")
+              if (converged) {
+                mes <- "Reached convergence.                 "
+                cat(mes, "\r")
+                cat("\n")
+              }
+              flush.console()
+            }
+            
+            break
+          }
+          
+          else if (j == n_lambda_w) {
+            index_lw <- 1
+            tmp_lw <- coglasso_obj$lambda_w[[index_lw]]
+            if (tmp_lw == lw_sel) converged <- TRUE
+            
+            lw_sel <- tmp_lw
+            
+            if (j == 1) {
+              coglasso_obj$opt_variability[i] <- variability_tmp
+              if (!light) {
+                coglasso_obj$merge_lw[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+              }
+            }
+            
+            if (verbose) {
+              mes <- "Conducting Subsampling Lambda_w....done.                 "
+              cat(mes, "\r")
+              cat("\n")
+              if (converged) {
+                mes <- "Reached convergence.                 "
+                cat(mes, "\r")
+                cat("\n")
+              }
+              flush.console()
+            }
+            
+            break
+          }
+          
+          coglasso_obj$opt_variability[i] <- variability_tmp
+          if (!light) {
+            coglasso_obj$merge_lw[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+          }
+        }
+      } 
+      else {
+        if (old_sampling) {
+          corr_matrixes <- vector(mode = "list", length = rep_num)
+          for (j in 1:rep_num) {
+            ind.sample <- sample(c(1:n), floor(n * stars_subsample_ratio), replace = FALSE)
+            corr_matrixes[[j]] <- cor(scale(coglasso_obj$data[ind.sample, ]))
+          }
+        }
+        
+        for (j in 1:n_lambda_b) {
+          if (verbose) {
+            mes <- paste(c("Conducting Subsampling Lambda_b....in progress:", floor(100 * j / n_lambda_b), "%"), collapse = "")
+            cat(mes, "\r")
+            flush.console()
+          }
+          real_rep.num <- 0
+          merge_tmp <- rep(0, d*d)
+          for (k in 1:rep_num) {
+            tmp <- co_glasso(corr_matrixes[[k]], pX, t(as.matrix(c(alpha, lw_sel, coglasso_obj$lambda_b[j]))), FALSE, FALSE, FALSE)
+            convergence <- tmp$convergence[1]
+            tmp <- as.vector(tmp$path[[1]])
+            if (convergence == 1) {
+              real_rep.num <- real_rep.num + 1
+              merge_tmp <- merge_tmp + tmp
+            }
+          }
+          rm(tmp)
+          gc()
+          
+          merge_tmp <- merge_tmp / real_rep.num
+          variability_tmp <- 4 * sum(merge_tmp * (1 - merge_tmp)) / (d * (d - 1))
+          
+          if (variability_tmp >= stars_thresh) {
+            index_lb <- max(j - 1, 1)
+            tmp_lb <- coglasso_obj$lambda_b[[index_lb]]
+            if (tmp_lb == lb_sel) converged <- TRUE
+            
+            lb_sel <- tmp_lb
+            
+            if (j == 1) {
+              coglasso_obj$opt_variability[i] <- variability_tmp
+              if (!light) {
+                coglasso_obj$merge_lb[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+              }
+            }
+            
+            if (verbose) {
+              mes <- "Conducting Subsampling Lambda_b....done.                 "
+              cat(mes, "\r")
+              cat("\n")
+              if (converged) {
+                mes <- "Reached convergence.                 "
+                cat(mes, "\r")
+                cat("\n")
+              }
+              flush.console()
+            }
+            
+            break
+          }
+          
+          else if (j == n_lambda_b) {
+            index_lb <- 1
+            tmp_lb <- coglasso_obj$lambda_b[[index_lb]]
+            if (tmp_lb == lb_sel) converged <- TRUE
+            
+            lb_sel <- tmp_lb
+            
+            if (j == 1) {
+              coglasso_obj$opt_variability[i] <- variability_tmp
+              if (!light) {
+                coglasso_obj$merge_lb[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+              }
+            }
+            
+            if (verbose) {
+              mes <- "Conducting Subsampling Lambda_b....done.                 "
+              cat(mes, "\r")
+              cat("\n")
+              if (converged) {
+                mes <- "Reached convergence.                 "
+                cat(mes, "\r")
+                cat("\n")
+              }
+              flush.console()
+            }
+            
+            break
+          }
+          
+          coglasso_obj$opt_variability[i] <- variability_tmp
+          if (!light) {
+            coglasso_obj$merge_lb[[i]] <- merge_tmp #make sure to make it a matrix again if implementing the vector to compute variability above
+          }
+        }
+      }
+      
+      select_lw <- !select_lw
+      
+      num_iter <- num_iter + 0.5
+    }
+    if (num_iter == max_iter) {
+      if (verbose) {
+        mes <- "Reached max Iterations.                 "
+        cat(mes, "\r")
+        cat("\n")
+        flush.console()
+      }
+    }
+    
+    if (verbose) {
+      mes <- "Selecting best Lambda_w/Lambda_b combination for all c values with \"xestars\"....done"
       cat(mes, "\r")
       flush.console()
     }
@@ -307,12 +719,12 @@ xstars <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NUL
 #' @examples
 #' cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
 #' \donttest{
-#' # Deprecated, use xstars() instead. Takes around 20 seconds
+#' # Deprecated, use xstars() instead. Takes around one minute
 #' sel_cg <- stars_coglasso(cg, verbose = FALSE)
 #' # ->
 #' sel_cg <- xstars(cg, verbose = FALSE)
 #' }
-stars_coglasso <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NULL, rep_num = 20, max_iter = 10, verbose = TRUE) {
-  lifecycle::deprecate_warn("1.1.0", "stars_coglasso()", "xstars()")
-  xstars(coglasso_obj, stars_thresh, stars_subsample_ratio, rep_num, max_iter, verbose)
+stars_coglasso <- function(coglasso_obj, stars_thresh = 0.1, stars_subsample_ratio = NULL, rep_num = 20, max_iter = 10, verbose = FALSE) {
+  lifecycle::deprecate_warn("1.1.0", "stars_coglasso()", "xstars()", always = TRUE)
+  xstars(coglasso_obj, stars_thresh, stars_subsample_ratio, rep_num, max_iter, verbose = FALSE)
 }