+++ R/coglasso.R
@@ -83,8 +83,8 @@
 #' # Typical usage: set the number of hyperparameters to explore
 #' cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
 #' \donttest{
-#' # Model selection using eXtended StARS, takes around 20 seconds
-#' sel_cg_xstars <- select_coglasso(cg, method = "xstars", verbose = FALSE)
+#' # Model selection using eXtended Efficient StARS, takes less than five seconds
+#' sel_cg_xestars <- select_coglasso(cg, method = "xestars", verbose = FALSE)
 #' }
 #' 
 coglasso <- function(data, pX, lambda_w = NULL, lambda_b = NULL, c = NULL, nlambda_w = NULL, nlambda_b = NULL, nc = NULL, lambda_w_max = NULL, lambda_b_max = NULL, c_max = NULL, lambda_w_min_ratio = NULL, lambda_b_min_ratio = NULL, c_min_ratio = NULL, cov_output = FALSE, verbose = TRUE) {